requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-telegram-bot==20.7
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
telethon==1.33.1
